datacite plugin - change file name and delete old
