+++ test/datacite_process_test.py
@@ -8,9 +8,10 @@ import csv
 import json
 
 class DataciteProcessTest(unittest.TestCase):
+
     def setUp(self) -> None:
-        self.test_dir = os.path.join('test', 'datacite_process')
-        self.zst_input_folder = os.path.join(self.test_dir, 'sample_dc')
+        self.test_dir = 'datacite_process'
+        self.json_dir = os.path.join(self.test_dir, 'jsonFiles')
         self.output_dir = os.path.join(self.test_dir, 'output_dir')
         self.processing_test_dir = os.path.join('test', 'datacite_processing')
         self.publisher_mapping = os.path.join(self.processing_test_dir, 'publishers.csv')
@@ -25,78 +26,12 @@ class DataciteProcessTest(unittest.TestCase):
         self.bad_dir = os.path.join(self.output_dir, '_bad')
         self.citations_output_path = self.output_dir + "_citations"
 
-    def test_malformed_line_report_saved_and_cleaned(self):
-        """
-        Verifica che con un NDJSON malformato venga creato il report degli errori:
-        - esiste output_dir/_bad/datacite_test_bad.ndjson.bad.ndjson
-        - contiene almeno una riga e include '10.1234/bad'
-        Poi elimina file e cartelle create, incluso _bad.
-        """
-        # PRE-CLEAN: rimuovi decompressioni residue nella cartella errori
-        if os.path.exists(self.error_input_folder):
-            for el in os.listdir(self.error_input_folder):
-                if el.endswith("decompr_zst_dir"):
-                    shutil.rmtree(os.path.join(self.error_input_folder, el))
-
-        # PRE-CLEAN: output, citations e bad
-        if os.path.exists(self.citations_output_path):
-            shutil.rmtree(self.citations_output_path)
-        if os.path.exists(self.output_dir):
-            shutil.rmtree(self.output_dir)
-        if os.path.exists(self.bad_dir):
-            shutil.rmtree(self.bad_dir)
-        if os.path.exists(self.db):
-            os.remove(self.db)
-        if os.path.exists(self.cache):
-            os.remove(self.cache)
-
-        # ESECUZIONE
-        preprocess(
-            datacite_ndjson_dir=self.error_input_folder,
-            publishers_filepath=self.publisher_mapping,
-            orcid_doi_filepath=self.iod,
-            csv_dir=self.output_dir,
-            redis_storage_manager=False,
-            storage_path=self.db,
-            cache=self.cache,
-            use_orcid_api=True
-        )
-
-        # VERIFICHE
-        bad_report_file = os.path.join(self.bad_dir, "datacite_test_bad.ndjson.bad.ndjson")
-        self.assertTrue(os.path.exists(bad_report_file), "Report errori non generato")
-        with open(bad_report_file, "r", encoding="utf-8") as bf:
-            content = bf.read()
-        self.assertTrue(len(content) > 0, "Report errori vuoto")
-        self.assertIn("10.1234/bad", content, "Report non contiene la riga malformata attesa")
-
-        # POST-CLEAN: rimuovi tutto, incluso _bad e file report
-        if os.path.exists(self.citations_output_path):
-            shutil.rmtree(self.citations_output_path)
-        if os.path.exists(self.output_dir):
-            shutil.rmtree(self.output_dir)
-        if os.path.exists(self.bad_dir):
-            shutil.rmtree(self.bad_dir)
-        if os.path.exists(self.db):
-            os.remove(self.db)
-        if os.path.exists(self.cache):
-            os.remove(self.cache)
-        # rimuovi eventuali decompressioni residue
-        if os.path.exists(self.error_input_folder):
-            for el in os.listdir(self.error_input_folder):
-                if el.endswith("decompr_zst_dir"):
-                    shutil.rmtree(os.path.join(self.error_input_folder, el))
-
     def test_preprocess_base_decompress_and_read(self):
         """Test base functionalities of the Datacite processor for producing META csv tables and INDEX tables:
-        1) All the files in the zst in input are correctly processed
+        1) All the files in input dir are correctly processed
         2) The number of files in input corresponds to the number of files in output for citations
         3) The number of files in input are duplicated in the output folder for both citing and cited entities
         """
-
-        for el in os.listdir(self.zst_input_folder):
-            if el.endswith("decompr_zst_dir"):
-                shutil.rmtree(os.path.join(self.zst_input_folder, el))
 
         if os.path.exists(self.output_dir):
             shutil.rmtree(self.output_dir)
@@ -110,7 +45,12 @@ class DataciteProcessTest(unittest.TestCase):
         if os.path.exists(bad_dir):
             shutil.rmtree(bad_dir)
 
-        preprocess(datacite_ndjson_dir=self.zst_input_folder, publishers_filepath=self.publisher_mapping,
+        if os.path.exists(self.db):
+            os.remove(self.db)
+        if os.path.exists(self.cache):
+            os.remove(self.cache)
+
+        preprocess(datacite_json_dir=self.json_dir, publishers_filepath=self.publisher_mapping,
                    orcid_doi_filepath=self.iod, csv_dir=self.output_dir, redis_storage_manager=False,
                    storage_path=self.db, cache=self.cache)
 
@@ -121,9 +61,11 @@ class DataciteProcessTest(unittest.TestCase):
                 cits_rows = list(csv.DictReader(f))
                 citations_in_output += len(cits_rows)
 
-        expected_citations_in_output = 4
+        #one self citation must not be considered
+        expected_citations_in_output = 19
 
-        expected_entities_in_output = 7
+        #excluding duplicated entities and one invalid doi 10.46979/rbn.v52i4.5546
+        expected_entities_in_output = 22
 
         self.assertEqual(expected_citations_in_output, citations_in_output)
 
@@ -138,7 +80,7 @@ class DataciteProcessTest(unittest.TestCase):
 
         self.assertEqual(expected_entities_in_output, entities_in_meta_output)
 
-        input_files_n = 1
+        input_files_n = 2
         self.assertTrue(citations_files_n == input_files_n)
 
         # CLEAN: output, _bad, decompressioni e db
@@ -147,9 +89,6 @@ class DataciteProcessTest(unittest.TestCase):
         bad_dir = os.path.join(self.output_dir, '_bad')
         if os.path.exists(bad_dir):
             shutil.rmtree(bad_dir)
-        for el in os.listdir(self.zst_input_folder):
-            if el.endswith("decompr_zst_dir"):
-                shutil.rmtree(os.path.join(self.zst_input_folder, el))
         if os.path.exists(self.db):
             os.remove(self.db)
         if os.path.exists(self.cache):
@@ -161,17 +100,18 @@ class DataciteProcessTest(unittest.TestCase):
         ORCIDs must not appear in _subject.csv files.
         """
         # Pre-clean
-        for el in os.listdir(self.zst_input_folder):
-            if el.endswith("decompr_zst_dir"):
-                shutil.rmtree(os.path.join(self.zst_input_folder, el))
         if os.path.exists(self.output_dir):
             shutil.rmtree(self.output_dir)
+
         citations_output_path = self.output_dir + "_citations"
         if os.path.exists(citations_output_path):
             shutil.rmtree(citations_output_path)
+
+        # assicura corretto funzionamento di _bad
         bad_dir = os.path.join(self.output_dir, '_bad')
         if os.path.exists(bad_dir):
             shutil.rmtree(bad_dir)
+
         if os.path.exists(self.db):
             os.remove(self.db)
         if os.path.exists(self.cache):
@@ -179,7 +119,7 @@ class DataciteProcessTest(unittest.TestCase):
 
         # Run with API disabled and no index
         preprocess(
-            datacite_ndjson_dir=self.zst_input_folder,
+            datacite_json_dir=self.json_dir,
             publishers_filepath=self.publisher_mapping,
             orcid_doi_filepath=None,
             csv_dir=self.output_dir,
@@ -204,38 +144,9 @@ class DataciteProcessTest(unittest.TestCase):
         self.assertFalse(found_orcid)
 
         # Post-clean
-        if os.path.exists(citations_output_path):
-            shutil.rmtree(citations_output_path)
+        # CLEAN: output, _bad, decompressioni e db
         if os.path.exists(self.output_dir):
             shutil.rmtree(self.output_dir)
-        bad_dir = os.path.join(self.output_dir, '_bad')
-        if os.path.exists(bad_dir):
-            shutil.rmtree(bad_dir)
-        for el in os.listdir(self.zst_input_folder):
-            if el.endswith("decompr_zst_dir"):
-                shutil.rmtree(os.path.join(self.zst_input_folder, el))
-        if os.path.exists(self.db):
-            os.remove(self.db)
-        if os.path.exists(self.cache):
-            os.remove(self.cache)
-
-    def test_preprocess_base_decompress_and_read_redis(self):
-        """Test base functionalities of the Datacite processor for producing META csv tables and INDEX tables:
-        1) All the files in the zst in input are correctly processed
-        2) The number of files in input corresponds to the number of files in output for citations
-        3) The number of files in input are duplicated in the output folder for both citing and cited entities
-        """
-
-        for el in os.listdir(self.zst_input_folder):
-            if el.endswith("decompr_zst_dir"):
-                shutil.rmtree(os.path.join(self.zst_input_folder, el))
-
-        if os.path.exists(self.output_dir):
-            shutil.rmtree(self.output_dir)
-
-        citations_output_path = self.output_dir + "_citations"
-        if os.path.exists(citations_output_path):
-            shutil.rmtree(citations_output_path)
         bad_dir = os.path.join(self.output_dir, '_bad')
         if os.path.exists(bad_dir):
             shutil.rmtree(bad_dir)
@@ -244,70 +155,25 @@ class DataciteProcessTest(unittest.TestCase):
         if os.path.exists(self.cache):
             os.remove(self.cache)
 
-        preprocess(datacite_ndjson_dir=self.zst_input_folder, publishers_filepath=self.publisher_mapping,
-                   orcid_doi_filepath=self.iod, csv_dir=self.output_dir, redis_storage_manager=True,
-                   storage_path=self.db, cache=self.cache)
-
-        citations_in_output = 0
-
-        for file in os.listdir(citations_output_path):
-            with open(os.path.join(citations_output_path, file), 'r', encoding='utf-8') as f:
-                cits_rows = list(csv.DictReader(f))
-                citations_in_output += len(cits_rows)
-
-        expected_citations_in_output = 4
-
-        expected_entities_in_output = 7
-
-        self.assertEqual(expected_citations_in_output, citations_in_output)
-
-        citations_files_n = len(list(os.listdir(citations_output_path)))
-
-        shutil.rmtree(citations_output_path)
-
-        entities_in_meta_output = 0
-        for file in os.listdir(self.output_dir):
-            with open(os.path.join(self.output_dir, file), 'r', encoding='utf-8') as f:
-                entities_in_meta_output += len(list(csv.DictReader(f)))
-
-        self.assertEqual(expected_entities_in_output, entities_in_meta_output)
-
-        input_files_n = 1
-        self.assertTrue(citations_files_n == input_files_n)
-
-        # CLEAN
-        if os.path.exists(self.output_dir):
-            shutil.rmtree(self.output_dir)
-        bad_dir = os.path.join(self.output_dir, '_bad')
-        if os.path.exists(bad_dir):
-            shutil.rmtree(bad_dir)
-        for el in os.listdir(self.zst_input_folder):
-            if el.endswith("decompr_zst_dir"):
-                shutil.rmtree(os.path.join(self.zst_input_folder, el))
-        if os.path.exists(self.db):
-            os.remove(self.db)
-        if os.path.exists(self.cache):
-            os.remove(self.cache)
 
     def test_preprocess_orcid_api_disabled_no_leak(self):
         """With ORCID API disabled, authors should not contain [orcid:] unless the DOI is in the provided index.
         Our sample input DOIs with authors having ORCID nameIdentifiers are not covered by the sample index (iod),
         so no [orcid:] should appear in the subject CSVs."""
 
-        # Clean any previous decompression dirs
-        for el in os.listdir(self.zst_input_folder):
-            if el.endswith("decompr_zst_dir"):
-                shutil.rmtree(os.path.join(self.zst_input_folder, el))
-
-        # Clean outputs
+        # Pre-clean
         if os.path.exists(self.output_dir):
             shutil.rmtree(self.output_dir)
+
         citations_output_path = self.output_dir + "_citations"
         if os.path.exists(citations_output_path):
             shutil.rmtree(citations_output_path)
+
+        # assicura corretto funzionamento di _bad
         bad_dir = os.path.join(self.output_dir, '_bad')
         if os.path.exists(bad_dir):
             shutil.rmtree(bad_dir)
+
         if os.path.exists(self.db):
             os.remove(self.db)
         if os.path.exists(self.cache):
@@ -315,7 +181,7 @@ class DataciteProcessTest(unittest.TestCase):
 
         # Run the process with ORCID API disabled
         preprocess(
-            datacite_ndjson_dir=self.zst_input_folder,
+            datacite_json_dir=self.json_dir,
             publishers_filepath=self.publisher_mapping,
             orcid_doi_filepath=self.iod,
             csv_dir=self.output_dir,
@@ -340,21 +206,38 @@ class DataciteProcessTest(unittest.TestCase):
         self.assertGreater(subject_rows, 0)
         self.assertEqual(orcid_mentions, 0)
 
-        # Cleanup
-        shutil.rmtree(citations_output_path)
-        shutil.rmtree(self.output_dir)
+        # Post-clean
+        # CLEAN: output, _bad, decompressioni e db
+        if os.path.exists(self.output_dir):
+            shutil.rmtree(self.output_dir)
         bad_dir = os.path.join(self.output_dir, '_bad')
         if os.path.exists(bad_dir):
             shutil.rmtree(bad_dir)
-        for el in os.listdir(self.zst_input_folder):
-            if el.endswith("decompr_zst_dir"):
-                shutil.rmtree(os.path.join(self.zst_input_folder, el))
         if os.path.exists(self.db):
             os.remove(self.db)
         if os.path.exists(self.cache):
             os.remove(self.cache)
 
     def test_any_db_creation_redis_no_testing(self):
+
+        # Pre-clean
+        if os.path.exists(self.output_dir):
+            shutil.rmtree(self.output_dir)
+
+        citations_output_path = self.output_dir + "_citations"
+        if os.path.exists(citations_output_path):
+            shutil.rmtree(citations_output_path)
+
+        # assicura corretto funzionamento di _bad
+        bad_dir = os.path.join(self.output_dir, '_bad')
+        if os.path.exists(bad_dir):
+            shutil.rmtree(bad_dir)
+
+        if os.path.exists(self.db):
+            os.remove(self.db)
+        if os.path.exists(self.cache):
+            os.remove(self.cache)
+
         try:
             rsm = RedisStorageManager(testing=False)
             rsm.set_value("TEST VALUE", False)
@@ -366,61 +249,55 @@ class DataciteProcessTest(unittest.TestCase):
         if run_test:
             rsm.del_value("TEST VALUE")
             if not len(rsm.get_all_keys()):
-                preprocess(datacite_ndjson_dir=self.zst_input_folder, publishers_filepath=self.publisher_mapping,
+                preprocess(datacite_json_dir=self.json_dir, publishers_filepath=self.publisher_mapping,
                            orcid_doi_filepath=self.iod, csv_dir=self.output_dir, redis_storage_manager=True,
                            storage_path=self.db, cache=self.cache)
 
-                for el in os.listdir(self.zst_input_folder):
-                    if el.endswith("decompr_zst_dir"):
-                        shutil.rmtree(os.path.join(self.zst_input_folder, el))
                 rsm.delete_storage()
 
             else:
 
                 print("test skipped: 'test_storage_management_no_testing' because redis db 2 is not empty")
+
+        # Post-clean
+        # CLEAN: output, _bad, decompressioni e db
+        if os.path.exists(self.output_dir):
+            shutil.rmtree(self.output_dir)
+        bad_dir = os.path.join(self.output_dir, '_bad')
+        if os.path.exists(bad_dir):
+            shutil.rmtree(bad_dir)
         if os.path.exists(self.db):
             os.remove(self.db)
         if os.path.exists(self.cache):
             os.remove(self.cache)
-        # rimuovi sempre eventuale _bad e output
-        bad_dir = os.path.join(self.output_dir, '_bad')
-        if os.path.exists(self.citations_output_path):
-            shutil.rmtree(self.citations_output_path)
-        if os.path.exists(self.output_dir):
-            shutil.rmtree(self.output_dir)
-        if os.path.exists(bad_dir):
-            shutil.rmtree(bad_dir)
-        for el in os.listdir(self.zst_input_folder):
-            if el.endswith("decompr_zst_dir"):
-                shutil.rmtree(os.path.join(self.zst_input_folder, el))
 
     def test_cache(self):
         'Nothing should be produced in output, since the cache file reports that all the files in input were completed'
 
-        for el in os.listdir(self.zst_input_folder):
-            if el.endswith("decompr_zst_dir"):
-                shutil.rmtree(os.path.join(self.zst_input_folder, el))
-
+        # Pre-clean
         if os.path.exists(self.output_dir):
             shutil.rmtree(self.output_dir)
 
         citations_output_path = self.output_dir + "_citations"
         if os.path.exists(citations_output_path):
             shutil.rmtree(citations_output_path)
+
+        # assicura corretto funzionamento di _bad
         bad_dir = os.path.join(self.output_dir, '_bad')
         if os.path.exists(bad_dir):
             shutil.rmtree(bad_dir)
+
         if os.path.exists(self.db):
             os.remove(self.db)
-        if os.path.exists(self.cache_test):
-            os.remove(self.cache_test)
+        if os.path.exists(self.cache):
+            os.remove(self.cache)
 
         with open(self.cache_test, "w", encoding="utf-8") as write_cache:
-            processed_files_dict = {'first_iteration': ['chunk_1'],
-                                    'second_iteration': ['chunk_1']}
+            processed_files_dict = {'first_iteration': ['jSonFile_1', 'jSonFile_2'],
+                                    'second_iteration': ['jSonFile_1', 'jSonFile_2']}
             json.dump(processed_files_dict, write_cache)
 
-        preprocess(datacite_ndjson_dir=self.zst_input_folder, publishers_filepath=self.publisher_mapping,
+        preprocess(datacite_json_dir=self.json_dir, publishers_filepath=self.publisher_mapping,
                    orcid_doi_filepath=self.iod, csv_dir=self.output_dir, redis_storage_manager=False,
                    storage_path=self.db, cache=self.cache_test)
 
@@ -450,21 +327,17 @@ class DataciteProcessTest(unittest.TestCase):
         self.assertEqual(expected_entities_in_output, unique_entities)
         self.assertEqual(expected_citations_in_output, citations_in_output)
 
-        # CLEAN
-        shutil.rmtree(citations_output_path)
-        shutil.rmtree(self.output_dir)
+        # Post-clean
+        # CLEAN: output, _bad, decompressioni e db
+        if os.path.exists(self.output_dir):
+            shutil.rmtree(self.output_dir)
         bad_dir = os.path.join(self.output_dir, '_bad')
         if os.path.exists(bad_dir):
             shutil.rmtree(bad_dir)
-
-        for el in os.listdir(self.zst_input_folder):
-            if el.endswith("decompr_zst_dir"):
-                shutil.rmtree(os.path.join(self.zst_input_folder, el))
-
         if os.path.exists(self.db):
             os.remove(self.db)
-        if os.path.exists(self.cache_test):
-            os.remove(self.cache_test)
+        if os.path.exists(self.cache):
+            os.remove(self.cache)
 
 if __name__ == '__main__':
     unittest.main()
